rename: relocate component files
