+++ src/components/Stock/SelectStock.tsx
